add kam

+++ src/ccnl-core/src/ccnl-dump.c
@@ -216,7 +216,7 @@ ccnl_dump(int lev, int typ, void *p)
         case CCNL_PACKET:
             INDENT(lev);
             CONSOLE("%p PACKET %s typ=%llu cont=%p contlen=%zd finalBI=%lld flags=0x%04x\n",
-                    (void *) pkt, ccnl_suite2str(pkt->suite), pkt->type,
+                    (void *) pkt, ccnl_suite2str(pkt->suite), (unsigned long long) pkt->type,
                     (void *) pkt->content, pkt->contlen, pkt->val.final_block_id,
                     pkt->flags);
             ccnl_dump(lev + 1, CCNL_PREFIX, pkt->pfx);
